Add an option to return to the other packageformat if it exists from the Snap/Packagepage Fixes #561
